# Conflicts:
#	frontend/viewer/src/project/SyncDialog.svelte

+++ frontend/viewer/src/home/HomeView.svelte
@@ -6,10 +6,8 @@
   import storybookIcon from '../stories/assets/storybook-icon.svg';
   import DevContent, {isDev} from '$lib/layout/DevContent.svelte';
   import {
-    useFwLiteConfig,
     useImportFwdataService,
     useProjectsService,
-    useTroubleshootingService,
   } from '$lib/services/service-provider';
   import TroubleshootDialog from '$lib/troubleshoot/TroubleshootDialog.svelte';
   import ServersList from './ServersList.svelte';
@@ -28,10 +26,12 @@
   import {cubicOut} from 'svelte/easing';
   import {transitionContext} from './transitions';
   import Anchor from '$lib/components/ui/anchor/anchor.svelte';
+  import FeedbackDialog from '$lib/about/FeedbackDialog.svelte';
+  import DeleteDialog from '$lib/entry-editor/DeleteDialog.svelte';
+  import {SYNC_DIALOG_QUERY_PARAM} from '../project/SyncDialog.svelte';
 
   const projectsService = useProjectsService();
   const importFwdataService = useImportFwdataService();
-  const fwLiteConfig = useFwLiteConfig();
   const exampleProjectName = 'Example-Project';
   const [send, receive] = crossfade({
     duration: 500,
@@ -45,9 +45,9 @@
       .replace(/-$/, '');
   }
 
-  let customExampleProjectName: string = '';
+  let customExampleProjectName = $state('');
 
-  let createProjectLoading = false;
+  let createProjectLoading = $state(false);
 
   async function createExampleProject() {
     try {
@@ -67,10 +67,16 @@
     }
   }
 
-  let deletingProject: undefined | string = undefined;
+  let deletingProject = $state<string>();
 
   async function deleteProject(project: IProjectModel) {
+    if (!deleteDialog) throw new Error('Delete dialog not initialized');
     try {
+      const syncDialogUrl = `/project/${project.code}/browse?${SYNC_DIALOG_QUERY_PARAM}=true`;
+      if (!await deleteDialog.prompt($t`Project`, $t`${project.name}`, {
+        isDangerous: true,
+        details: $t`Make sure your [changes are synced](${syncDialogUrl}) to Lexbox.`,
+      })) return;
       deletingProject = project.id;
       await projectsService.deleteProject(project.code);
       await refreshProjects();
@@ -79,7 +85,7 @@
     }
   }
 
-  let importing = '';
+  let importing = $state('');
 
   async function importFwDataProject(name: string) {
     if (importing) return;
@@ -92,9 +98,9 @@
     }
   }
 
-  let projectsPromise = projectsService
+  let projectsPromise = $state(projectsService
     .localProjects()
-    .then((projects) => projects.sort((p1, p2) => p1.name.localeCompare(p2.name)));
+    .then((projects) => projects.sort((p1, p2) => p1.name.localeCompare(p2.name))));
 
   async function refreshProjects() {
     let promise = projectsService.localProjects().then((p) => p.sort((p1, p2) => p1.name.localeCompare(p2.name)));
@@ -102,8 +108,7 @@
     projectsPromise = promise;
   }
 
-  const supportsTroubleshooting = useTroubleshootingService();
-  let troubleshootDialog: TroubleshootDialog | undefined;
+  let troubleshootDialog = $state<TroubleshootDialog>();
 
   let clickCount = 0;
   let clickTimeout: ReturnType<typeof setTimeout> | undefined;
@@ -126,12 +131,17 @@
     }
   }
 
+  let feedbackOpen = $state(false);
+
+  let deleteDialog = $state<DeleteDialog>();
 </script>
+
+<DeleteDialog bind:this={deleteDialog}/>
 
 <AppBar tabTitle={$t`Dictionaries`}>
   {#snippet title()}
     <div class="text-lg flex gap-2 items-center">
-      <Icon onclick={clickIcon} src={mode.current === 'dark' ? logoLight : logoDark} alt={$t`Lexbox logo`}/>
+      <Icon onclick={clickIcon} src={mode.current === 'dark' ? logoLight : logoDark} class="size-8" alt={$t`Lexbox logo`}/>
       <h3>{$t`Dictionaries`}</h3>
     </div>
   {/snippet}
@@ -150,21 +160,18 @@
       <ResponsiveMenu.Root>
         <ResponsiveMenu.Trigger/>
         <ResponsiveMenu.Content>
-          <ResponsiveMenu.Item href={fwLiteConfig.feedbackUrl} target="_blank" icon="i-mdi-chat-question">
+          <ResponsiveMenu.Item onSelect={() => feedbackOpen = true} icon="i-mdi-chat-question">
             {$t`Feedback`}
           </ResponsiveMenu.Item>
-          {#if supportsTroubleshooting}
-            <ResponsiveMenu.Item
-              icon="i-mdi-face-agent"
-              onSelect={() => troubleshootDialog?.open()}>
-              {$t`Troubleshoot`}
-            </ResponsiveMenu.Item>
-          {/if}
+          <ResponsiveMenu.Item
+            icon="i-mdi-face-agent"
+            onSelect={() => troubleshootDialog?.open()}>
+            {$t`Troubleshoot`}
+          </ResponsiveMenu.Item>
         </ResponsiveMenu.Content>
       </ResponsiveMenu.Root>
-      {#if supportsTroubleshooting}
-        <TroubleshootDialog bind:this={troubleshootDialog}/>
-      {/if}
+      <FeedbackDialog bind:open={feedbackOpen}/>
+      <TroubleshootDialog bind:this={troubleshootDialog}/>
     </div>
     {/snippet}
 </AppBar>
@@ -188,35 +195,37 @@
                     onclick={() => refreshProjects()}/>
           </div>
           <div>
-            {#each projects.filter((p) => p.crdt) as project, i (project.id ?? i)}
+            {#each projects.filter((p) => p.crdt) as project (project.id ?? project)}
               {@const server = project.server}
               {@const loading = deletingProject === project.id}
               <div out:send={{key: 'project-' + project.code}} in:receive={{key: 'project-' + project.code}}>
-                <Anchor href={`/project/${project.code}`}>
-                  <ProjectListItem icon="i-mdi-book-edit-outline"
-                                   {project}
-                                   {loading}
-                                   subtitle={!server ? $t`Local only` : $t`Synced with ${server.displayName}`}
-                  >
-                    {#snippet actions()}
-                      <div class="flex items-center">
-                        {#if $isDev}
-                          <Button
-                            icon="i-mdi-delete"
-                            variant="ghost"
-                            title={$t`Delete`}
-                            class="p-2 hover:bg-primary/20"
-                            onclick={(e) => {
-                            e.preventDefault();
-                            void deleteProject(project);
-                          }}
-                          />
-                        {/if}
-                        <Icon icon="i-mdi-chevron-right" class="p-2"/>
-                      </div>
+                <ResponsiveMenu.Root contextMenu>
+                  <ResponsiveMenu.Trigger>
+                    {#snippet child({props})}
+                      <Anchor {...props} href={`/project/${project.code}`}>
+                        <ProjectListItem icon="i-mdi-book-edit-outline"
+                                         {project}
+                                         {loading}
+                                         subtitle={!server ? $t`Local only` : $t`Synced with ${server.displayName}`}
+                        >
+                          {#snippet actions()}
+                            <div class="flex items-center">
+                              <Icon icon="i-mdi-chevron-right" class="p-2"/>
+                            </div>
+                          {/snippet}
+                        </ProjectListItem>
+                      </Anchor>
                     {/snippet}
-                  </ProjectListItem>
-                </Anchor>
+                  </ResponsiveMenu.Trigger>
+                  <ResponsiveMenu.Content>
+                    <ResponsiveMenu.Item icon="i-mdi-bug" onSelect={() => troubleshootDialog?.open(project.code)}>
+                      {$t`Troubleshoot`}
+                    </ResponsiveMenu.Item>
+                    <ResponsiveMenu.Item icon="i-mdi-delete" onSelect={() => void deleteProject(project)}>
+                      {$t`Delete`}
+                    </ResponsiveMenu.Item>
+                  </ResponsiveMenu.Content>
+                </ResponsiveMenu.Root>
               </div>
             {/each}
             <DevContent>
@@ -229,7 +238,7 @@
               </Anchor>
             </DevContent>
             {#if !projects.some(p => p.name === exampleProjectName) || $isDev}
-              <ListItem onclick={() => createExampleProject()} disabled={createProjectLoading} class="dark:bg-muted/50 bg-muted/80 hover:bg-muted/30 hover:dark:bg-muted">
+              <ListItem onclick={() => createExampleProject()} loading={createProjectLoading}>
                 <span>{$t`Create Example Project`}</span>
                 {#snippet actions()}
                   <div class="flex flex-nowrap items-center gap-2">
@@ -253,7 +262,7 @@
           <div>
             <p class="sub-title">{$t`Classic FieldWorks Projects`}</p>
             <div>
-              {#each projects.filter((p) => p.fwdata) as project (project.id ?? project.name)}
+              {#each projects.filter((p) => p.fwdata) as project (project.name)}
                 <Anchor href={`/fwdata/${project.code}`}>
                   <ProjectListItem {project}>
                     {#snippet icon()}
